Refactor solvers to classes + tidying + fix logging
* Refactor solvers into classes for clearer code reuse and encapsulation.
* Separate `MeshDecorator` into `MeshIters` and `MeshLoader`.
* Create/use `MeshDumper` for managing mesh file dumping, both intermediate and final result (via RAII).
* Fix logging of Fastor tensors by ensuring `operator<<` is in `Fastor` namespace.
* Attempt to reduce unnecessary links to Boost static libs by using `find_package` targets rather than catch-all `CONAN_PKG` target.

+++ src/lib/FeltElements/internal/Attributes.hpp
@@ -150,7 +150,7 @@ protected:
 		const auto to_tensor = [this, &cell_vtxhs](BoundaryElement::VtxhIdxs const & face_idxs) {
 			BoundaryElement::Vtxhs face_vtxhs;
 			for (auto const & [face_idx, cell_idx] : boost::adaptors::index(face_idxs))
-				face_vtxhs[face_idx] = cell_vtxhs[cell_idx];
+				face_vtxhs[static_cast<std::size_t>(face_idx)] = cell_vtxhs[cell_idx];
 			return for_element(face_vtxhs);
 		};
